Refactor constructor
